requirements.txt
openpyxl>=3.1.0
python-docx>=1.1.0
pywebview>=5.0.0
pyinstaller>=6.0.0
